Add navbar + UI improvements About is given its own page with the math section Better structure Player analysis and Party Analysis Small tweaks to forms for better usability
